Split Rent and Return View & Added success popup
split rent into a rent and return view and fixed controller for that. redirected rents and return to go to the home page after. implemented pop up with success message. need to work on failure and pop up style. need to work on closing pop up by not clicking on it.  Also condensed the report view